UX improvements
* Add placeholder prop
* Prevent propagating enter keypress if we handled it inside component.
Still propagates if we returned early. This allows for most intuitive
user experience where if we selected something from the suggestion box
or entered the text manually enter keypress will accept the change, and
when we have no input enter will propagate upwards so it can be handled
upstream e.g. for accepting the changes.